requirements.txt
discord.py==2.0.0
py-dactyl==2.0.0